Update SchoolLunchCS302.cpp
This part of the code creates a loop that keeps asking the user whether they would like to add another item to their order. The user is prompted with the message "Would you like to add another item? Y/N :". The user's input is then stored in the variable addItem. If the user enters anything other than 'Y' or 'y', the loop is exited using the 'break' statement. This allows the program to proceed to the next step in the code.